+++ cli/CHANGELOG.md
@@ -5,6 +5,7 @@ _Released 2/13/2024 (PENDING)_
 
 **Bugfixes:**
 
+- Fixed an issue which caused the browser to relaunch after closing the browser from the Launchpad. Fixes [#28852](https://github.com/cypress-io/cypress/issues/28852).
 - Fixed incorrect type definition of `dom.getContainsSelector`. Fixed in [#28339](https://github.com/cypress-io/cypress/pull/28339).
 
 **Dependency Updates:**